clean up tests + rename
