+++ core/mock_test.go
@@ -6,14 +6,31 @@ import (
 	rpctest "github.com/celestiaorg/celestia-core/rpc/test"
 )
 
-// StartMockNode starts a mock core node background process and returns it.
+// StartMockNode starts a mock Core node background process and returns it.
 func StartMockNode() *node.Node {
 	app := kvstore.NewApplication()
 	app.RetainBlocks = 10
 	return rpctest.StartTendermint(app, rpctest.SuppressStdout, rpctest.RecreateConfig)
 }
 
-// MockClient returns a started Mock Core Client.
-func MockClient() Client {
+// MockEmbeddedClient returns a started mock Core Client.
+func MockEmbeddedClient() Client {
 	return NewEmbeddedFromNode(StartMockNode())
+}
+
+// StartRemoteClient returns a started mock Core Client
+// from the remote Core node process.
+func StartRemoteClient() (*node.Node, Client, error) {
+	remote := StartMockNode()
+	protocol, ip := getRemoteEndpoint(remote)
+	client, err := NewRemote(protocol, ip)
+	return remote, client, err
+}
+
+// getRemoteEndpoint returns the protocol and ip of the remote node.
+func getRemoteEndpoint(remote *node.Node) (string, string) {
+	endpoint := remote.Config().RPC.ListenAddress
+	// protocol = "tcp"
+	protocol, ip := endpoint[:3], endpoint[6:]
+	return protocol, ip
 }
